* fixed CMSG_CAST_SPELL structure, actualized SpellCastTargetFlags enum

+++ src/Client/World/CMSGConstructor.cpp
@@ -117,7 +117,7 @@ void WorldSession::SendCastSpell(uint32 spellid, bool nocheck)
 
     WorldPacket packet;
     ByteBuffer temp;
-    uint16 flags=TARGET_FLAG_SELF; // target mask. spellcast implementeation needs to be changed if TARGET_MASK_SELF is != 0
+    uint32 flags=TARGET_FLAG_SELF; // target mask. spellcast implementeation needs to be changed if TARGET_MASK_SELF is != 0
     packet << spellid;
     packet << (uint8)0; // unk
     if(target && my->GetTarget() != GetGuid()) // self cast?
